working through agent changes
